Refactor: rename function file for clarity
